WIP Extract product loading into its own renderless component
There's still some issue with the recursive loading with this new
implementation and it goes into an infinite loop. It's a cleaner separation
now, though. It's still a question if the products should actually be stored
in the same store or not. They are big and that might make using Redux dev
tooling painful.